requirements.txt
aiofiles==0.8.0
anyio==3.6.1
asgiref==3.5.1
base58==2.1.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.77.1
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
# ipfshttpclient==0.7.0
jose==1.0.0
msgpack==1.0.3
multiaddr==0.0.9
netaddr==0.8.0
numpy==1.22.3
pandas==1.4.2
passlib==1.7.4
py-algorand-sdk==1.13.1
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
pydantic==1.9.0
pymongo==4.1.1
PyNaCl==1.5.0
PyPDF2==2.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
varint==1.0.2

ipfsapi~=0.4.4
pyteal~=0.10.1